Add shell variable functions.
This is probably the last commit I will do.
Because I find it both not educational and boring from now on.
The remaining works are:
- Improve parser rules of form VAR=VALUE.
- Implement syntax parser.
- Maintain return code.
- Connect IO redirection to commands.

And...I find C really verbose to use. The problem is that in C, you have to
build lots of basic blocks for your program to really work.

Maybe I should have used glibc I suppose.

Or...just adopting to C++ may be a better choice for me.
I don't know. C++ is also verbose anyway.